intital

+++ header/huffman.h
@@ -2,8 +2,8 @@
 #define HUFFMAN
 
 typedef struct n { 
-	float poid ; 
-  int caractere ;
+	int poid ; 
+	int caractere ;
 	noeud* fils_gauche ; 
 	noeud* fils_droite ; 
 	noeud* pere ;
